lucy_ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
openai
httpx
solana
langchain
langchain-openai
langchain-google-genai
solders
python-jose[cryptography]